Removed vars from generator and replaced it with auto-generated defaults.
The previous approach required the user to manually choose defaults for
the variables used in the build edges. Since the only purpose is to
reduce the size of ninja build files, this is rather an implementation
detail. Therefore I've decided to hide that from the user and instead
automatically pick the most used value for each variable as the
default.