Disallow changes to BUILDDIR in Python.
The problem is that both the ninjapie shell script and the Python code
needs to know where the build directory is. Since the Python code is
called from the shell script after the build directory was already used,
the Python code cannot change it.

For that reason, I've decided to introduce the NPBUILD environment
variable that specifies the build directory and initializes env
['BUILDDIR'] from it. Changes to this variable are disallowed.
Furthermore, other build-directory related variables have been removed
now or are implicitly relative to env['BUILDDIR'].