Reworked the globbing concept and rewrote main script in Python.
The idea is to record the used glob patterns and check only them for
added/removed files. This has the advantage that we know where to
actually search for files and also allows the user decide whether he
wants to use it. If glob isn't used, we will also not search through
the file system anymore. Therefore, the user can choose between more
convenience and faster build times.

I've also rewritten the main script in Python to avoid a second language
and use exactly the same glob function for checking.