Use per-environment unique names for generated object files.
The rational is that we should be able to build the same source file in
different environments without interference. If the output file is
chosen by the user, that's fine. But if the output file is generated by
Ninjapie (like for object files by default), we should make sure that
the names don't clash.

Therefore, each environment object now has a unique id and we append the
id to the name of the object file.