Let the user install libraries and completely control the LIBPATH.
It may be more convenient that self-build libraries are magically
installed and found during linking, but it causes more confusion and
trouble than it helps. For example, there might be name clashes if
multiple libraries have the same name or a library has the same name as
an existing library on the host.

Therefore, I think it is less error prone to simply let the user decide
where libraries are installed to (if at all) and how to find them.