Added --version argument.
This is actually a bit tricky, because we want to be able to use
ninjapie as an installed package and when just having it somewhere in
our file system (e.g., cloned from Github). Therefore we have a
version.py with the version number now and import that in
the __init__.py for packaged usages, but read and exec the version.py
from the __main__.py for packaged and unpackaged usages.